+++ pkg/transformers/senml/message.go
