+++ readers/api/endpoint_test.go
@@ -10,10 +10,10 @@ import (
 	"testing"
 
 	"github.com/mainflux/mainflux"
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/readers"
 	"github.com/mainflux/mainflux/readers/api"
 	"github.com/mainflux/mainflux/readers/mocks"
-	"github.com/mainflux/mainflux/transformers/senml"
 	"github.com/stretchr/testify/assert"
 )
