+++ readers/cassandra/messages.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/gocql/gocql"
 	"github.com/mainflux/mainflux/pkg/errors"
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/readers"
-	"github.com/mainflux/mainflux/transformers/senml"
 )
 
 var errReadMessages = errors.New("faled to read messages from cassandra database")
