+++ readers/cassandra/messages_test.go
@@ -8,9 +8,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/readers"
 	creaders "github.com/mainflux/mainflux/readers/cassandra"
-	"github.com/mainflux/mainflux/transformers/senml"
 	cwriters "github.com/mainflux/mainflux/writers/cassandra"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
