+++ readers/mongodb/messages_test.go
@@ -10,9 +10,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/readers"
 	mreaders "github.com/mainflux/mainflux/readers/mongodb"
-	"github.com/mainflux/mainflux/transformers/senml"
 	mwriters "github.com/mainflux/mainflux/writers/mongodb"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
