+++ readers/postgres/messages.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/jmoiron/sqlx" // required for DB access
 	"github.com/mainflux/mainflux/pkg/errors"
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/readers"
-	"github.com/mainflux/mainflux/transformers/senml"
 )
 
 const errInvalid = "invalid_text_representation"
