+++ readers/postgres/messages_test.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/gofrs/uuid"
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/readers"
 	preader "github.com/mainflux/mainflux/readers/postgres"
-	"github.com/mainflux/mainflux/transformers/senml"
 	pwriter "github.com/mainflux/mainflux/writers/postgres"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
