+++ writers/cassandra/messages.go
@@ -6,7 +6,7 @@ package cassandra
 import (
 	"github.com/gocql/gocql"
 	"github.com/mainflux/mainflux/pkg/errors"
-	"github.com/mainflux/mainflux/transformers/senml"
+	"github.com/mainflux/mainflux/pkg/transformers/senml"
 	"github.com/mainflux/mainflux/writers"
 )
